Cleanup view management, use ModelForms for creating programs and subplans (#79)
* Clean up form submission, removing HTTP request in middle

* Manually verify name/year pairs in new degree submission code

* Migrate views into individual files, clean up URLs

* Migrate programs and subplans to using proper Django forms
