+++ cassdegrees/templates/createsubplan.html
@@ -6,51 +6,35 @@
 
 {% block content %}
 
-    {% if msg != None %}
-        <p class="{% if is_error %}msg-error{% else %}msg-success{% endif %}">{{ msg }}</p>
+    {% if submitted %}
+        <p class="msg-success">Program successfully added!</p>
     {% endif %}
 
-    <form class="anuform" action="/create_subplan/" method="post">
+    {% for error in form.non_field_errors %}
+         <p class="msg-error">{{ error }}</p>
+    {% endfor %}
 
+    <form class="anuform" action="" method="post">
         {% csrf_token %}
 
         <fieldset>
-
             <legend>Subplan details</legend>
 
-            <!-- As HTML forms only support get or post methods,
-            add this hidden input for consistency with the rest of the forms -->
-            <input type="hidden" name="_method" value="post">
+            {{ form.management_form }}
 
-            <p>
-                <label for="code">Subplan Code:</label>
-                <input class="text" id="code" type="text" name="code" placeholder="e.g. ANTH-MIN"
-                       aria-required="true" required value="{{ code|default:"" }}">
-            </p>
-
-            <p>
-                <label for="year">Subplan Year:</label>
-                <input class="text" id="year" type="number" name="year"
-                       aria-required="true" required min="2000" value="{{ year|default:"2019" }}">
-            </p>
-
-            <p>
-                <label for="name">Subplan Name:</label>
-                <input class="text" id="name" type="text" name="name"
-                       aria-required="true" placeholder="e.g. Anthropology" required value="{{ name|default:"" }}">
-            </p>
-
-            <p>
-                <label for="plantype">Subplan Type:</label>
-                <select class="text" id="planType" type="text" name="planType"
-                        aria-required="true" required">
-                    <option value="MAJ" {% if planType == "MAJ" %}selected{% endif %}>Major</option>
-                    <option value="MIN" {% if planType == "MIN" %}selected{% endif %}>Minor</option>
-                    <option value="SPEC" {% if planType == "SPEC" %}selected{% endif %}>Specialisation</option>
-                </select>
-            </p>
-
+            {% for field in form %}
+                <p class="form-group">
+                    {% if not field.is_hidden %}
+                        {{ field.label_tag }}
+                    {% endif %}
+                    {{ field }}
+                </p>
+                {% for error in field.errors %}
+                     <div class="msg-error inline-error">{{ error }}</div>
+                {% endfor %}
+            {% endfor %}
         </fieldset>
+
         <p class="text-right">
             <input class="btn-uni-grad btn-large" type ="submit" value="Create">
         </p>
